requirements_cleaned.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.6.2
func-timeout==4.3.5
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
numpy==2.0.0
openai==1.34.0
pip==24.0
pydantic==2.7.4
pydantic-core==2.18.4
pymysql==1.1.1
readline
setuptools==69.5.1
sniffio==1.3.1
tqdm==4.66.4
typing-extensions==4.12.2
tzdata
wheel==0.43.0
faiss-cpu
transformers
datasets
pandas
sqlalchemy
sqlite-utils
bitsandbytes
accelerate
peft
trl